2048
This is my implementation of the classic 2048 game using C++ and the SFML library. Immerse yourself in a visually engaging gaming experience as you slide numbered tiles on the grid, aiming to merge and reach the coveted 2048 tile. Explore the codeand challenge yourself to achieve the highest score in this rendition of the popular 2048 game.